+++ sdk/healthinsights/azure-healthinsights-radiologyinsights/azure/healthinsights/radiologyinsights/aio/__init__.py
@@ -6,18 +6,11 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._client import RadiologyInsightsClient
-
-try:
-    from ._patch import __all__ as _patch_all
-    from ._patch import *  # pylint: disable=unused-wildcard-import
-except ImportError:
-    _patch_all = []
+from ._patch import RadiologyInsightsClient
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
     "RadiologyInsightsClient",
 ]
-__all__.extend([p for p in _patch_all if p not in __all__])
 
 _patch_sdk()
